requirements.txt
asgiref==3.9.2
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==4.2
django-allauth==65.11.2
django-crispy-forms==2.4
django-smart-selects==1.7.2
django-widget-tweaks==1.5.0
idna==3.10
moncashify==1.1.0
mysqlclient==2.2.7
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
pypdfium2==4.30.0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
reportlab==4.4.3
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
